set blade color on BLADE_REFRESH
thanks to blooper on the swgediscord.com discord server for pointing this one out. when a BLADE_REFRESH command is received, the blade color wasn't being set. however some color changing legacy hilts like Ahsoka and Cal Kestis use BLADE_REFRESH to change colors. so we do need to set the color when a refresh is received.